上传文件

+++ src/main/java/com/hklk/oplatform/controller/school/EditTeacherController.java
@@ -32,10 +32,11 @@ public class EditTeacherController extends BaseController {
     /**
      * 2018/7/4 16:09
      * 查询老师
-     * @param param     筛选条件
-     * @param pageNum   分页参数
-     * @author 曹良峰
+     *
+     * @param param   筛选条件
+     * @param pageNum 分页参数
      * @return java.lang.String
+     * @author 曹良峰
      */
     @ResponseBody
     @RequestMapping("/queryTeachers")
@@ -48,9 +49,10 @@ public class EditTeacherController extends BaseController {
     /**
      * 2018/7/4 16:10
      * 根据条件查询老师
-     * @param sTeacher  老师对象
-     * @author 曹良峰
+     *
+     * @param sTeacher 老师对象
      * @return java.lang.String
+     * @author 曹良峰
      */
     @ResponseBody
     @RequestMapping("/selectTeacherByParam")
@@ -64,9 +66,10 @@ public class EditTeacherController extends BaseController {
     /**
      * 2018/7/4 16:10
      * 添加或修改老师
+     *
      * @param sTeacher 老师对象
-     * @author 曹良峰
      * @return code ：1009 老师已存在  200 成功
+     * @author 曹良峰
      */
     @ResponseBody
     @RequestMapping("/addOrUpdateTeacher")
@@ -74,12 +77,14 @@ public class EditTeacherController extends BaseController {
                                      HttpServletResponse response, HttpSession session) {
         Integer schoolId = getLoginSchool(request).getSchoolId();
         STeacher param = new STeacher();
+
         param.setPhone(sTeacher.getPhone());
         param.setSchoolId(schoolId);
+
         STeacher result = sTeacherService.selectBySTeacher(param);
         if (sTeacher.getId() == null && result != null) {
             return ToolUtil.buildResultStr(StatusCode.TEACHER_EX, StatusCode.getStatusMsg(StatusCode.TEACHER_EX));
-        } else if (sTeacher.getId() != null && result != null && result.getId() != sTeacher.getId()) {
+        } else if (sTeacher.getId() != null && result != null && result.getId().intValue() != sTeacher.getId().intValue()) {
             return ToolUtil.buildResultStr(StatusCode.TEACHER_EX, StatusCode.getStatusMsg(StatusCode.TEACHER_EX));
         } else {
             sTeacher.setSchoolId(schoolId);
@@ -87,14 +92,15 @@ public class EditTeacherController extends BaseController {
             return ToolUtil.buildResultStr(StatusCode.SUCCESS, StatusCode.getStatusMsg(StatusCode.SUCCESS));
         }
     }
-    
+
     /**
      * 2018/7/4 16:12
      * 修改老师状态
-     * @param id    老师id
-     * @param status    状态
-     * @author 曹良峰
+     *
+     * @param id     老师id
+     * @param status 状态
      * @return java.lang.String
+     * @author 曹良峰
      */
     @ResponseBody
     @RequestMapping("/updateTeacherStatus")
@@ -110,9 +116,10 @@ public class EditTeacherController extends BaseController {
     /**
      * 2018/7/4 16:13
      * 删除老师
-     * @param id    老师id
-     * @author 曹良峰
+     *
+     * @param id 老师id
      * @return java.lang.String
+     * @author 曹良峰
      */
     @ResponseBody
     @RequestMapping("/delTeacher")
